+++ src/decorators/class.decorator.ts
@@ -1,5 +1,9 @@
 import { createServiceMap, serviceMap } from './utilities';
-import { AxiosfitInterceptor } from '../interfaces';
+import { AxiosfitInterceptor, AxiosfitRequestInterceptor, AxiosfitResponseInterceptor, AxiosfitConfig } from '../interfaces';
+
+const defaultAxiosfitConfig: AxiosfitConfig = {
+  usePromises: false,
+};
 
 /**
  * Class decorator.
@@ -7,13 +11,18 @@ import { AxiosfitInterceptor } from '../interfaces';
  *
  * @param {string} [endpointPath] Optional. Common endpoint for all the methods inside the class.
  */
-export function HTTP(endpointPath?: string) {
+export function HTTP(endpointPath?: string | AxiosfitConfig, axiosfitConfig?: AxiosfitConfig) {
   return constructor => {
     const serviceName = constructor.serviceName || constructor.name;
+    const ep = endpointPath && typeof endpointPath === 'string' ? endpointPath : '';
+    if (typeof endpointPath === 'object') {
+      axiosfitConfig = endpointPath;
+    }
     if (!serviceMap[serviceName]) {
       createServiceMap(constructor);
     }
-    serviceMap[serviceName].setBaseServiceEndpoint(endpointPath || '');
+    serviceMap[serviceName].setBaseServiceEndpoint(ep);
+    serviceMap[serviceName].setAxiosfitConfig(axiosfitConfig || defaultAxiosfitConfig);
   };
 }
 
@@ -21,10 +30,11 @@ export function HTTP(endpointPath?: string) {
  * Class decorator.
  * Decorator to be used to add a global request interceptor.
  *
- * @param {AxiosfitInterceptor[]} interceptors The interceptors.
+ * @param {(AxiosfitInterceptor | AxiosfitRequestInterceptor | AxiosfitResponseInterceptor)[]} interceptors The interceptors.
  */
-export function Interceptors<T = AxiosfitInterceptor>(...interceptors: T[]) {
+export function Interceptors(...interceptors: any[]) {
   return constructor => {
+    // TODO: Ensure interceptors are one of them
     const serviceName = constructor.serviceName || constructor.name;
     if (!serviceMap[serviceName]) {
       createServiceMap(constructor);