+++ test/services/promises/TestPromiseServiceNoBase.ts
@@ -1,7 +1,7 @@
 import { HTTP, GET, AxiosResponse } from '../../../src';
 import { TestRoutes } from '../TestRoutes';
 
-@HTTP()
+@HTTP({ usePromises: true })
 export class TestPromiseServiceNoBase {
   @GET(TestRoutes.GET.REQUEST.URL)
   public performGetRequest(): Promise<AxiosResponse<string>> {